+++ lizard_progress/static/lizard_progress/download_page.js
@@ -1,5 +1,0 @@
-// The download page can also contain upload dialogs, they call the
-// upload_page_functions.refresh_uploaded_file_tables() functions when
-// upload is complete.
-var upload_page_functions = (function () {
-})();