+++ lizard_progress/views/__init__.py
@@ -3,3 +3,4 @@ from lizard_progress.views.upload import *
 from lizard_progress.views.views import *
 from lizard_progress.views.wizards import *
 from lizard_progress.views.configuration import *
+from lizard_progress.views.users import *